AuthCookie: made more Py3 compliant
The load() method of http.BaseCookie has been change substantial in Python 3.
Unfortunately this is not documented in the online docs, the only hint I found,
was a comment in the http.cookie module::

  We first parse the whole cookie string and reject it if it's
  syntactically invalid (this helps avoid some classes of injection
  attacks).

This means, cookies with a leading 'Set-Cookie:' string like::

  'Set-Cookie: _M2AUTH_="exp=1485714802.7772.... "'

are no longer valid for BaseCookie.load() and will **silently** result in an
unset/empty Cookie. Here is a small Py3 example::

    >>> from http.cookies import SimpleCookie
    >>> C1 = SimpleCookie()
    >>> C1["foo"] = "bar"
    >>> C2.output()
    'Set-Cookie: foo=bar'
    >>> C2 = SimpleCookie()
    >>> C2.load(C.output())
    >>> C2.output()
    ''

Loaded cookie is unset, see empty string on last line of the example above.

   This also affects the AuthCookie.isGoodCookieString()
   method, which uses BaseCookie.load() to validate a cookie!!

The workaround is to use an empty header argument when using outpout()::

    >>> C2.load(C.output(header=""))
    >>> C2.output()
   'Set-Cookie: foo=bar'

To get in use of this workaround, AuthCookie.output() is fixed in a way it is
more BaseCookie compliant.

BTW: AuthCookieJar.isGoodCookieString returned ambiguous bool or int values,
which is also fixed within in this patch (return always a bool value).

Signed-off-by: Markus Heiser <markus.heiser@darmarit.de>